feat: add types to board repo
